requirements.txt
django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
numpy==1.23.3
pandas==2.0.1
scikit-learn==1.2.2
# tensorflow==2.12.0
pillow==9.5.0
python-dotenv==1.0.0
